Throw exception if no text or data segment is declared
If the first statement isn't a .text or .data directive, the Parser
now throws an exception. If main isn't declared in a text segment,
the Parser throws an exception.
Fixed problems with Memory about the boundaries of text and data
segments when an typical data directive is in a text segment